requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
joblib==1.5.2
numpy==2.3.5
pandas==2.3.3
protobuf==6.33.1
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
wheel==0.45.1
